requirements.txt
# Core Framework
fastapi>=0.109.0,<0.110.0
uvicorn[standard]>=0.27.0,<0.28.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# AI APIs
cohere>=5.20.0,<6.0.0
openai>=1.3.0,<2.0.0
anthropic>=0.7.0,<1.0.0
openrouter>=0.1.0

# Vector DB
qdrant-client>=1.7.0,<2.0.0

# Database
sqlalchemy>=2.0.0,<3.0.0
alembic>=1.13.0,<2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0

# HTTP
httpx>=0.25.0,<1.0.0
requests>=2.31.0,<3.0.0

# Logging
structlog>=23.2.0
python-json-logger>=2.0.0
